backend/tests/test_analyzer.py
import pytest
from services.video_analyzer import (
    _title_performance_score,
    _pacing_score,
    _tag_quality_score,
    _keyword_difficulty,
    _upload_timing_score
)
from services import analyzer_config as C

def test_title_performance_ideal():
    # Ideal title: 40-60 chars, power word, number
    title = "Top 10 Ways to Code Better Python Fast Today!" # 45 chars
    # Base 50 + Length 20 + Power(Top, Fast, Today -> 21) + Number 9 = 100
    score = _title_performance_score(title)
    assert score == 100

def test_title_performance_short():
    title = "Hi"
    # Base 50 - Short 15 = 35.
    assert _title_performance_score(title) == 35

def test_pacing_score():
    # 8-20 mins = 480-1200 sec -> Score 90
    assert _pacing_score(600) == 90
    # 2 mins = 120 sec -> Score 50
    assert _pacing_score(120) == 50

def test_tag_quality():
    # 0 tags
    assert _tag_quality_score([]) == C.TAG_QUALITY["BASELINE"]
    
    # Good mix
    tags = ["python", "python tutorial", "how to learn python for beginners"]
    # Count: 3 * 4 = 12.
    # Diversity: Short(1) + Medium(2 words) + Long(6 words) = 3 -> 3 * 15 = 45.
    # Total: 12 + 45 = 57.
    # Validated against min(Count, 50) -> min(12, 50) = 12.
    # Penalty: too_long (>30 chars) = 1 * 5 = 5.
    # Total: 57 - 5 = 52.
    assert _tag_quality_score(tags) == 52

def test_keyword_difficulty():
    title = "How to Code"
    # "how to" is high comp.
    # Base 55 + 8 = 63.
    val, label = _keyword_difficulty(title)
    assert val == 63
    assert label == "Medium" # 40 <= 63 < 70

def test_upload_timing():
    # This test is time-sensitive and hard to verify without mocking.
    # Skipping for now to avoid flakes.
    pass
